Update apipkg to v3.0.1
